Fix bug with spawning items
Originally, if you were to spawn some items and then spawn a bunch more and letting them despawn, but keeping some items spawned, the counter would overflow and could end up using the same value as an item that is already spawned. This change should avoid that.

Technically, this code will still cause that to happen if the counter falls below 256, but it is highly unlikely that the code will be used that many times.