Added additional features
Added a display that allows the player to see the frame that they tried to guard/superguard. If they press the button early, then it'll say how many frames early, to a specific extent.

Added a cheat that allows the player to walk through most objects. They will not be able to walk through NPCs.