Bug Fix + Added Additional Features
Fixed the game crashing if the player entered las_08 with the Sequence at 385 and with GSW(1121) at 7. GSW(1121) will be reduced to 6 in this instance in order to avoid the crash.

Allowed being able to give actors badges inside of battles. When an actor that is neither Mario nor the partners is given a badge, all of the current badges equipped are cleared, and the new badge is equipped.

Adjusted the window length for the button timing display to be dynamic.