Small Bug Fixes and Changes
Fixed the adjustable value menu being shown for a frame after pressing A or B. This would also cause it to draw an incorrect amount of digits to be changed.

Changed several misnamed variables.

Added additional enums for values used in the adjustable menu.

Modified the Change By Id and Change By Icon menu options to automatically set the value/position to the currently-selected item.