Bug Fix + Modified the Auto Action Commands code
Fixed the code that displays guard/superguard timings to work properly when an attack cannot be superguarded, or if the debug badge is equipped. If the player tries to superguard an attack that cannot be superguarded, a message will be printed saying so.

Adjusted the Auto Action Commands code to guard/superguard more attacks. It now effectively works exactly the same as the Debug Badge.