Adjusted several uses of party functions
Changed the parameter for partyGetPtr to an int32_t.

Removed conditional checks in getPartnerPointer, getFollowerPointer, removePartnerFromOverworld, and removeFollowerFromOverworld. This is because the functions that they call, partyGetPtr and partyKill2, already do the same conditional checks.

Removed a conditional check in menufunctions.cpp before calling removeFollowerFromOverworld, for the same reason listed above.

Used the proper variable type for partyEntry2Hello in menufunctions.cpp.