Stored the Link Register when calling functions via assembly
This is mainly done so that the backtrace screen will show the proper address(es) in the LR Save field when the function that was called at that point was done via injected assembly.

Doing this also allows using bl branches instead of regular branches, which actually ends up using less memory overall.

Renamed and/or moved some functions that are called via assembly.

Created a trampoline for the marioKeyOn fix, as it is easier to manage.

Adjusted the Player struct to include the key variable.