Adjusted the On-Screen Timer code, and added a Frame Counter code
The Original On-Screen Timer code has been renamed to Frame Counter, as it only counts the frames that are drawn.

The new On-Screen Timer code is based off of OSTime, so it will be accurate even when going through loading zones. When both timers are active at the same time, the Frame Counter will be placed above the On-Screen Timer.

Also made the following adjustments:
1. drawChangeButtonCombo now takes a button combo by pointer instead of by reference.
2. drawOnScreenTimerButtonCombos takes an array of button combos, as both the On-Screen Timer and Frame Counter use it.
3. Repositioned the OSLink functions in the lst files.
4. Removed the entry in setInitialSettings that sets the Lock Flags bool to false, as the bool is now unused.