Draw the float register values and the paired singles in the error handler
The error handler will now have 3 pages, with the float register values being drawn on page 2, and the paired singles being drawn on page 3. The main reason that they are on separate pages is because the game consistantly crashes when both the float register values and the paired singles are drawn at the same time, and the cause is still unknown.

Also changed the following:
1. The OSError variable holding the string for the type of OSError that occured is now stored in the ErrorHandlerStrings struct.

2. In the drawString auto function, the returned string from OSGetFontTexture is now used, the value returned is the pointer to the next character in the string. Therefore, the i variable in that function was removed.