Made several adjustments to the npcNameToPtr stuff
Adjusted the replacement function to simply call the original function, and then check if the returned pointer is valid.This change also fixed an issue where the function did not return the same pointer that the vanilla one would, as the vanilla one ends up going out of range when it fails to find a proper NPC.

Also changed the error text to display the number of times that the function returned an invalid pointer. Every time the number is incremented, the timer is reset to 5 seconds.