Small improvements to the current code for the Frame Advance cheat
The code is running from the start of makeKey, so the trampoline is now always called immediately.

Removed JustResumedGameplay variable, as it shouldn't be needed.

Use the initial value of AdvanceFrame to determine if the player is initially pausing or is currently unpausing.

In the infinite loop, the makeKey trampoline is now called at the end of the loop instead of the beginning, as it is already called immediately at the start of the function.