Set max widths for several things
Set it for the On-Screen Timer, the Frame Counter, and the Jump Storage displays.

drawTextMultipleLines now has a main function that takes the same params as drawTextMain, and it now calls that instead of drawText.

The code for getting the time in drawFrameCounter has been cleaned up to match the code in getTimeString.