Fix heap corruptions in pre-SQ phase 2 cutscenes
The first fix is for bom_01 (Fahr Outpost), and the second fix is for mri_00 (The Great Tree) (This one may not actually cause a heap corruption, but changing it anyway just to be safe).

Made several other adjustments to performRelPatches.

Added all of the rel ids to seq_mapchange.h.

Made checkIfInGame use the DMO enum from seq_mapchange.h.