Prevent animPoseMain from running if poseId is negative
If manually warping while on a chapter intro screen, the game will crash after all of the fade work has occurred, which is caused by -1 being passed into animPoseMain which occurs because the fades are resolved early via reloadRoomMain. Preventing animPoseMain from running in this case should avoid the issue.