Multiple displays adjustments
The default Y position for error text is now properly decremented when text from the heap corruption buffer is drawn, as it now accounts for how many lines are in the buffer.

Error texts are now drawn with differing orders to keep their drawn order consistent.

Changed the scale of error texts from 0.65 to 0.6.

Shortened some local pointer names.