Memory Editor - Change misc flags to state variable
Also prevented checking for button inputs on the same frame that the memory editor is opened, which would end up checking for one of the button combo buttons being pressed.

Also adjusted the drawing code to simply decrement by lineDecrement instead of doing unnecessary math to calculate the current line.