Remove various class drawing functions from classes and make them static
This is mainly because the drawing functions either use no variables from the class, or only use a small amount of them.

Also removed the scale variable from the Option Selector class.

Also made some small adjustments to the event warp code, so that the equipped badges are only backed up with the flag for it is set.

Also made some small adjustments in handleWarpByEvent and getEventDetails in warpsEvent.cpp.