Update REL Loader v3 codes and readme
Fixed issues with the game's crash screen occurring due to the loader not properly incrementing the stack before exiting.

Adjusted the loader to preserve the main function's parameters (The vanilla game doesn't use them, but it's possible that future mods might).

Adjusted the readme to clarify that v3 of the REL Loader uses fixed linking.

Adjusted the readme to clarify that mods can work with all REL Loader versions if support is added to the mods to allow for it.